# beat/docs/swagger.yaml
basePath: /api
definitions:
...
  presenters.MetadataErrorResponse:
    description: presenters.MetadataErrorResponse
    properties:
      data:
        example: ""
        type: string
      error:
        example: error message
        type: string
      status:
        example: false
        type: boolean
    type: object
  presenters.MetadataListResponse:
    description: presenters.MetadataListResponse
    properties:
      data: {}
      error:
        example: ""
        type: string
      status:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: Deals Beats, Beat_Listened, filtering
  title: Fiber Beat Service
  version: "1.0"
paths:
...
  /metadata/genres:
    get:
      description: Returns a list of all available music genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all genres
      tags:
      - Metadata
...
  /metadata/instruments:
    get:
      description: Returns a list of all available instruments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all instruments
      tags:
      - Metadata
...
  /metadata/moods:
    get:
      description: Returns a list of all available mood categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all moods
      tags:
      - Metadata
  /metadata/tags:
    get:
      description: Returns a list of all available tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all tags
      tags:
      - Tags
...
  /metadata/timestamps:
    get:
      description: Returns a list of all available timestamps
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all timestamps
      tags:
      - Timestamp
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"